moved sub-functions to relevant file, trial utility archiveFileFromCommand in src extsvc function
